refactor: switch #3885
