--- components/vc-switch/index.ts
@@ -1,4 +1,0 @@
-// base rc-switch 1.9.0
-import Switch from './src/Switch';
-
-export default Switch;
